windows-app/libre computer s905xccv1/requirements.txt
# Tiny Home Personal Assistant Requirements

# Core dependencies
speech-recognition>=3.10.0
pyttsx3>=2.90
requests>=2.31.0
asyncio-mqtt>=0.16.1
fastapi>=0.104.1
uvicorn>=0.24.0

# Audio processing
pyaudio>=0.2.11
sounddevice>=0.4.6
wave>=0.0.2

# Hardware control
RPi.GPIO>=0.7.1  # For GPIO if needed
pyserial>=3.5    # For LoRa communication

# Web interface
jinja2>=3.1.2
websockets>=12.0

# Optional AI/ML
openai>=1.3.0
transformers>=4.35.0
torch>=2.1.0

# Development tools
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0